+++ VacationRental.Api/Controllers/CalendarController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using VacationRental.Api.Models;
+using VacationRental.Api.Services;
 
 namespace VacationRental.Api.Controllers
 {
@@ -9,63 +10,18 @@ namespace VacationRental.Api.Controllers
     [ApiController]
     public class CalendarController : ControllerBase
     {
-        private readonly IDictionary<int, RentalViewModel> _rentals;
-        private readonly IDictionary<int, BookingViewModel> _bookings;
+        private readonly ICalendarService _calendarService;
 
-        public CalendarController(
-            IDictionary<int, RentalViewModel> rentals,
-            IDictionary<int, BookingViewModel> bookings)
+        public CalendarController(ICalendarService calendarService)
         {
-            _rentals = rentals;
-            _bookings = bookings;
+            _calendarService = calendarService;
         }
 
         [HttpGet]
         public CalendarViewModel Get(int rentalId, DateTime start, int nights)
         {
-            return GetCalendar(rentalId, start, nights);
+            return _calendarService.GetCalendar(rentalId, start, nights);
         }
 
-        private CalendarViewModel GetCalendar(int rentalId, DateTime start, int nights)
-        {
-            if (nights < 0)
-                throw new ApplicationException("Nights must be positive");
-            if (!_rentals.ContainsKey(rentalId))
-                throw new ApplicationException("Rental not found");
-
-            var result = new CalendarViewModel
-            {
-                RentalId = rentalId,
-                Dates = new List<CalendarDateViewModel>()
-            };
-            for (var i = 0; i < nights; i++)
-            {
-                var date = new CalendarDateViewModel
-                {
-                    Date = start.Date.AddDays(i),
-                    Bookings = new List<CalendarBookingViewModel>(),
-                    PreparationTimes = new List<CalendarPreparationTimeViewModel>()
-                };
-
-                foreach (var booking in _bookings.Values)
-                {
-                    if (booking.RentalId == rentalId
-                        && booking.Start <= date.Date && booking.End > date.Date)
-                    {
-                        date.Bookings.Add(new CalendarBookingViewModel { Id = booking.Id, Unit = booking.Unit });
-                    }
-
-                    if (booking.RentalId == rentalId
-                        && booking.End <= date.Date && booking.End.AddDays(_rentals[rentalId].PreparationTimeInDays) > date.Date)
-                    {
-                        date.PreparationTimes.Add(new CalendarPreparationTimeViewModel { Unit = booking.Unit });
-                    }
-                }
-
-                result.Dates.Add(date);
-            }
-
-            return result;
-        }
     }
 }
