+++ VacationRental.Api/Services/CalendarService.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using System.Collections.Generic;
+using VacationRental.Api.Models;
+
+namespace VacationRental.Api.Services
+{
+    public class CalendarService : ICalendarService
+    {
+        private readonly IDictionary<int, RentalViewModel> _rentals;
+        private readonly IDictionary<int, BookingViewModel> _bookings;
+
+        public CalendarService(
+            IDictionary<int, RentalViewModel> rentals,
+            IDictionary<int, BookingViewModel> bookings)
+        {
+            _rentals = rentals;
+            _bookings = bookings;
+        }
+
+        public CalendarViewModel GetCalendar(int rentalId, DateTime start, int nights)
+        {
+            if (nights < 0)
+                throw new ApplicationException("Nights must be positive");
+            if (!_rentals.ContainsKey(rentalId))
+                throw new ApplicationException("Rental not found");
+
+            var result = new CalendarViewModel
+            {
+                RentalId = rentalId,
+                Dates = new List<CalendarDateViewModel>()
+            };
+            for (var i = 0; i < nights; i++)
+            {
+                var date = new CalendarDateViewModel
+                {
+                    Date = start.Date.AddDays(i),
+                    Bookings = new List<CalendarBookingViewModel>(),
+                    PreparationTimes = new List<CalendarPreparationTimeViewModel>()
+                };
+
+                foreach (var booking in _bookings.Values)
+                {
+                    if (booking.RentalId == rentalId
+                        && booking.Start <= date.Date && booking.End > date.Date)
+                    {
+                        date.Bookings.Add(new CalendarBookingViewModel { Id = booking.Id, Unit = booking.Unit });
+                    }
+
+                    if (booking.RentalId == rentalId
+                        && booking.End <= date.Date && booking.End.AddDays(_rentals[rentalId].PreparationTimeInDays) > date.Date)
+                    {
+                        date.PreparationTimes.Add(new CalendarPreparationTimeViewModel { Unit = booking.Unit });
+                    }
+                }
+
+                result.Dates.Add(date);
+            }
+
+            return result;
+        }
+    }
+}
