+++ VacationRental.Api/Services/IBookingsService.cs
@@ -1,0 +1,10 @@
+﻿using VacationRental.Api.Models;
+
+namespace VacationRental.Api.Services
+{
+    public interface IBookingsService
+    {
+        ResourceIdViewModel AddBooking(BookingBindingModel model);
+        BookingViewModel GetBookingById(int bookingId);
+    }
+}
