+++ VacationRental.Api/Services/ICalendarService.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using VacationRental.Api.Models;
+
+namespace VacationRental.Api.Services
+{
+    public interface ICalendarService
+    {
+        CalendarViewModel GetCalendar(int rentalId, DateTime start, int nights);
+    }
+}
