+++ VacationRental.Api/Services/IRentalsService.cs
@@ -1,0 +1,11 @@
+﻿using VacationRental.Api.Models;
+
+namespace VacationRental.Api.Services
+{
+    public interface IRentalsService
+    {
+        ResourceIdViewModel CreateRental(RentalBindingModel model);
+        RentalViewModel GetRentalById(int rentalId);
+        void UpdateRental(int rentalId, RentalBindingModel model);
+    }
+}
