+++ VacationRental.Api/Startup.cs
@@ -25,8 +25,12 @@ namespace VacationRental.Api
 
             services.AddSwaggerGen(opts => opts.SwaggerDoc("v1", new Info { Title = "Vacation rental information", Version = "v1" }));
 
-            services.AddSingleton<IDictionary<int, RentalViewModel>>(new Dictionary<int, RentalViewModel>());
-            services.AddSingleton<IDictionary<int, BookingViewModel>>(new Dictionary<int, BookingViewModel>());
+            services.AddRentalDictionary();
+            services.AddBookingDictionary();
+
+            services.AddBookingsService();
+            services.AddCalendarService();
+            services.AddRentalsService();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
